Final code fixes/logic fixes, major fixes

+++ database/migrations/2020_01_25_182724_create_orders_table.php
@@ -21,7 +21,7 @@ class CreateOrdersTable extends Migration
             $table->unsignedBigInteger('order_group_id')->nullable();
             $table->json('ingredients')->nullable();
             $table->integer('quantity')->nullable()->default(1);
-            $table->integer('price')->nullable()->default(0);
+            $table->double('price', 8, 2)->nullable()->default(0);
             $table->timestamps();
             $table->softDeletes();
